requirements.txt
asgiref==3.11.0
attrs==25.4.0
black==25.11.0
Django==5.2.8
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
flake8==7.3.0
inflection==0.5.1
isort==7.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Markdown==3.10
mccabe==0.7.0
packaging==25.0
pillow==12.0.0
psycopg==3.2.13
psycopg-binary==3.2.13
pycodestyle==2.14.0
pyflakes==3.4.0
PyJWT==2.10.1
python-dotenv==1.2.1
pytokens==0.3.0
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.29.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
